to TRAVEL is to LIVE
This is fun virtual planner for your next trip to some of the most visited countries of the world : Italy, France, U.S.A. The form begins by taking personal details like name, DOB, country you live in and month that you plan to visit and further the choices regarding which country you plan your vacation, what kinds of places you like to visit (monuments, museums, malls) and what restaurants and cuisines you like. On submitting these details you get your personalised plan with weather details of the month, a link to book flights and hotels and various places to visit in that country according to your preference with their short description and history. It also gives images and locations of restaurants.